requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
pydantic==2.10.6
python-multipart==0.0.20
requests==2.31.0
langchain==0.3.21
pymongo==4.11.3
faiss-cpu==1.10.0
numpy==1.26.4
pandas==2.2.2
